+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - DSL.Parameter: `filterType()` of parameters. 
 - DSL.Parameter: `filterType()` of fields. 
 - dot: Remove need for specifying `"dot-ignore"` property.
+- DSL: re-registering the same entity using `entity(entitty, label...)` will now update the entity label.
 - CLI: `--edge-layout=spline|polyline|ortho`, used together with `-R`/`--render`.
 - `spring-axon`: member aggregate label changed to `$aggregate[$member]`.